Remove `set-output` usage in action
As `set-output` usage in Github Actions may be removed soon (postponed from 2023-07-24),
use writing to GITHUB_OUTPUT there as well.
It may introduce an error source if the GITHUB_OUTPUT file is unwritable (for unknown reasons).
Also print a message to stdout if an input is missing.

Maybe exit with an error as well when this happens?