ci: Fix third-party python deps
Maybe it would be good to fix security related python dependencies
as we don't, by default, trust third-party software.

This fixes jwt and also all of it's dependencies.

It hasn't been touched since 2021 so it may not be so critical.
Probably good practice.

Also, since it hasn't been touched since 2021 I guess we should
not have to worry about bumping versions too much.